RealEstateManager/app/src/main/java/com/sursulet/realestatemanager/data/local/Converters.kt
package com.sursulet.realestatemanager.data.local

import android.graphics.Bitmap
import android.graphics.BitmapFactory
import androidx.room.TypeConverter
import java.io.ByteArrayOutputStream
import java.time.LocalDate

class Converters {

    @TypeConverter
    fun fromString(value: String?): LocalDate? {
        return value?.let { LocalDate.parse(it) }
    }

    @TypeConverter
    fun dateToString(date: LocalDate?): String? = date.toString()

    @TypeConverter
    fun fromByteArray(value: ByteArray): Bitmap {
        return BitmapFactory.decodeByteArray(value,0,value.size)
    }

    @TypeConverter
    fun bitmapToByteArray(bitmap: Bitmap): ByteArray {
        val outputStream = ByteArrayOutputStream()
        bitmap.compress(Bitmap.CompressFormat.PNG,100,outputStream)
        return outputStream.toByteArray()
    }
}